Add example application of simplex
Given the nutritional information of a meal, and a list of ingredients
(with their own set of nutritional information) we want to calculate
the quantity (in grams) of each ingredient that satisfies the profile
of the meal.

We can solve with simplex by using as variables the quantity of the
ingredients which we then minimize given the sum of each nutrient
as constraints. I.e.:

Min. y = ingredient1+ ingredient2 + ... + ingredientN
Where:
   ingredient_calories1 + ... + ingredient_caloriesN = meal_calories
   ingredient_protein1 + ... + ingredient_proteinN = meal_protein
   etc...

The '=' constraint might be too restrictive. Given we don't need to
be accurate, we can choose a ε and convert each '=' constraint in a
pair of '>' and '<' constraints, i.e.:
  ingredient_nutrient1 + ... + ingredient_nutrient2 > meal_nutrient * (1 - ε)
  ingredient_nutrient1 + ... + ingredient_nutrient2 < meal_nutrient * (1 + ε)